requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
loguru==0.7.2
tqdm==4.66.1
PyYAML==6.0.1
pandas==2.1.1
psycopg2-binary==2.9.9